Finish all tasks related to onboarding
