+++ TaskSpaces/.idea/gradle.xml
@@ -15,5 +15,6 @@
         </option>
       </GradleProjectSettings>
     </option>
+    <option name="offlineMode" value="true" />
   </component>
 </project>
